requirements.txt
Django
mysqlclient
python-decouple
requests
Pillow
django-crispy-forms
crispy-bootstrap5
